Fix <AddDiscussionPost /> toggling
